Kore.ASTVerifier.AliasVerifier
